+++ lib/models/despesa.dart
@@ -1,35 +1,31 @@
-import 'dart:ffi';
-
 import 'package:aluguel/models/imovel.dart';
 
-class Despesa{
-
+class Despesa {
   String _servico;
-  Float _valor;
+  double _valor;
   Imovel _imovel;
 
-  String get servico{
+  String get servico {
     return this._servico;
   }
 
-  Float get valor{
+  double get valor {
     return this._valor;
   }
 
-  Imovel get imovel{
+  Imovel get imovel {
     return this._imovel;
   }
 
-  set servico(String servico){
+  set servico(String servico) {
     this._servico = servico;
   }
 
-  set valor(Float valor){
+  set valor(double valor) {
     this._valor = valor;
   }
 
-  set imovel(Imovel imovel){
+  set imovel(Imovel imovel) {
     this._imovel = imovel;
   }
-
 }
